utils update and project subcarpet remodeling
